+++ src/NowoIconSelectorBundle.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Nowo\IconSelectorBundle;
 
+use Nowo\IconSelectorBundle\DependencyInjection\Compiler\TwigPathsPass;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -16,4 +18,8 @@ use Symfony\Component\HttpKernel\Bundle\Bundle;
  */
 final class NowoIconSelectorBundle extends Bundle
 {
+    public function build(ContainerBuilder $container): void
+    {
+        $container->addCompilerPass(new TwigPathsPass());
+    }
 }